+++ fivefury/ydr/fixups/__init__.py
@@ -1,3 +1,3 @@
-from .drawable import audit_legacy_fragment_drawable_fixups
+from .drawable import audit_fragment_drawable_fixups
 
-__all__ = ["audit_legacy_fragment_drawable_fixups"]
+__all__ = ["audit_fragment_drawable_fixups"]